Langchain RAG using Async Fastapi and Qdrant/requirements.txt
langsmith
langchain-qdrant
langchain_core
aiomonitor
python-dotenv
mysql-connector-python
openai
fastapi
uvicorn
langchain
openai
langchain-openai
sentence-transformers
pypdf
nest_asyncio
IPython
PyPDF2
python-docx
beautifulsoup4
python-multipart
aiosqlite